+++ src/components/classDetail/embla.css
@@ -19,6 +19,7 @@
   touch-action: pan-y;
   height: 100%;
 }
+
 /* .embla__slide {
   flex: 0 0 var(--slide-size);
   min-width: 0;
@@ -116,11 +117,12 @@
   bottom: 0;
   left: 50%;
   transform: translateX(-50%);
+  border-radius: 10px;
 }
 
 .embla__dot--selected:after {
   box-shadow: inset 0 0 0 0.2rem var(--text-body);
-  border: 4px solid #6c5ff7;
+  border: 3.5px solid #6c5ff7;
 }
 
 .embla__dot {
@@ -142,80 +144,10 @@
   background-color: white;
   width: 100%;
   height: 6px;
-  border-radius: 2px;
+  border-radius: 6px;
   content: '';
 }
-
-/* .embla__dot.is-selected:after {
-  background-color: #1bcacd;
-  opacity: 1;
-  border: 3px solid black;
-} */
 
 .embla__button__svg {
   color: white;
 }
-
-/* .embla__button {
-  -webkit-tap-highlight-color: rgba(var(--text-high-contrast-rgb-value), 0.5);
-  -webkit-appearance: none;
-  appearance: none;
-  background-color: transparent;
-  touch-action: manipulation;
-  display: inline-flex;
-  text-decoration: none;
-  cursor: pointer;
-  border: 0;
-  padding: 0;
-  margin: 0;
-  box-shadow: inset 0 0 0 0.2rem var(--detail-medium-contrast);
-  width: 3.6rem;
-  height: 3.6rem;
-  z-index: 1;
-  border-radius: 50%;
-  color: var(--text-body);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.embla__button:disabled {
-  color: var(--detail-high-contrast);
-}
-
-.embla__dots {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: flex-end;
-  align-items: center;
-}
-.embla__dot {
-  -webkit-tap-highlight-color: rgba(var(--text-high-contrast-rgb-value), 0.5);
-  -webkit-appearance: none;
-  appearance: none;
-  background-color: transparent;
-  touch-action: manipulation;
-  display: inline-flex;
-  text-decoration: none;
-  cursor: pointer;
-  border: 0;
-  padding: 0;
-  margin: 0;
-  width: 2.6rem;
-  height: 2.6rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 50%;
-}
-.embla__dot:after {
-  box-shadow: inset 0 0 0 0.2rem var(--detail-medium-contrast);
-  border: 1px solid gray;
-  width: 1.4rem;
-  height: 1.4rem;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  content: '';
-}
-
-} */
